Fix: 修复启动闪退问题
- ColorPalette.fromArgbLong(): 使用 value.toInt() 替代直接使用 value
- StrokeRenderer.paintColor(): 使用 config.color.toInt() 替代直接使用 Long
- EditorScreen.paperColor: 使用 toInt() 转换
- HomeScreen 所有 Color(Long) 调用都改为 Color(Long.toInt())

这些修复解决了 Color(Long) 构造函数在某些情况下导致的崩溃问题

+++ app/src/main/java/com/xiaominote/app/drawing/ColorPalette.kt
@@ -44,5 +44,5 @@ object ColorPalette {
         return a or r or g or b
     }
 
-    fun fromArgbLong(value: Long): Color = Color(value)
+    fun fromArgbLong(value: Long): Color = Color(value.toInt())
 }
